Reorganize examples (#1784)
