+++ examples/control/ego_open_agent.py
@@ -6,10 +6,15 @@ NOTE: You will need to install Rust to run this example. See https://rustup.rs/ 
 
 import importlib
 import logging
+import sys
+from pathlib import Path
 
 import gym
 
 from smarts.core.utils.episodes import episodes
+
+sys.path.insert(0, str(Path(__file__).parents[1]))
+from tools.argument_parser import default_argument_parser
 
 try:
     open_agent = importlib.import_module("open_agent")
@@ -17,8 +22,6 @@ except ModuleNotFoundError as e:
     raise ModuleNotFoundError(
         f"Ensure that the open-agent has been installed with `pip install open-agent"
     )
-
-from argument_parser import default_argument_parser
 
 logging.basicConfig(level=logging.INFO)
