+++ examples/dynamic_history_vehicles_replacement.py
@@ -2,7 +2,7 @@ import logging
 import random
 from typing import Any, Callable, Dict, Sequence
 
-from argument_parser import default_argument_parser
+from tools.argument_parser import default_argument_parser
 
 from envision.client import Client as Envision
 from smarts.core import seed as random_seed
