+++ examples/trajectory_tracking_agent.py
@@ -1,7 +1,7 @@
 import logging
 
 import gym
-from argument_parser import default_argument_parser
+from tools.argument_parser import default_argument_parser
 
 from smarts.core.agent import Agent
 from smarts.core.agent_interface import AgentInterface, AgentType
